refactor: shell_ui.rsロジック1,107行→374行に分割 (native_menu/terms/splash/app_frame抽出)
