+++ crates/katana-ui/src/shell_ui.rs
@@ -12,15 +12,9 @@ use crate::app::*;
 
 use eframe::egui;
 
-use crate::{
-    app_state::{AppAction, ViewMode},
-    preview_pane::DownloadRequest,
-    widgets::StyledComboBox,
-};
+use crate::app_state::AppAction;
 
 const INVISIBLE_LABEL_SIZE: f32 = 0.1;
-/// Splash screen animation repaint interval (~30fps).
-const SPLASH_REPAINT_INTERVAL_MS: u64 = 32;
 
 pub(crate) fn invisible_label(text: &str) -> egui::RichText {
     egui::RichText::new(text)
@@ -72,173 +66,14 @@ pub(crate) fn indent_prefix(depth: usize) -> String {
 }
 
 // ─────────────────────────────────────────────
-// Split layout helpers
+// Native menu re-exports (implementation in native_menu.rs)
 // ─────────────────────────────────────────────
 
-// ─────────────────────────────────────────────
-// eframe::App implementation (egui main rendering loop)
-// ─────────────────────────────────────────────
-
-#[cfg(target_os = "macos")]
-mod native_menu {
-    // Must match the tag constants defined in macos_menu.m (Objective-C).
-    pub const TAG_OPEN_WORKSPACE: i32 = 1;
-    pub const TAG_SAVE: i32 = 2;
-    pub const TAG_LANG_EN: i32 = 3;
-    pub const TAG_LANG_JA: i32 = 4;
-    pub const TAG_ABOUT: i32 = 5;
-    pub const TAG_SETTINGS: i32 = 6;
-    pub const TAG_LANG_ZH_CN: i32 = 7;
-    pub const TAG_LANG_ZH_TW: i32 = 8;
-    pub const TAG_LANG_KO: i32 = 9;
-    pub const TAG_LANG_PT: i32 = 10;
-    pub const TAG_LANG_FR: i32 = 11;
-    pub const TAG_LANG_DE: i32 = 12;
-    pub const TAG_LANG_ES: i32 = 13;
-    pub const TAG_LANG_IT: i32 = 14;
-    pub const TAG_CHECK_UPDATES: i32 = 15;
-    pub const TAG_RELEASE_NOTES: i32 = 16;
-
-    // These FFI symbols are linked from Objective-C (macos_menu.m) and called
-    // only at runtime; the Rust compiler cannot see the call sites.
-    #[allow(dead_code)]
-    extern "C" {
-        pub fn katana_setup_native_menu();
-        pub fn katana_poll_menu_action() -> i32;
-        pub fn katana_set_app_icon_png(png_data: *const u8, png_len: std::ffi::c_ulong);
-        pub fn katana_set_process_name();
-        pub fn native_free_menu_actions();
-        pub fn katana_update_menu_strings(
-            file: *const std::ffi::c_char,
-            open_workspace: *const std::ffi::c_char,
-            save: *const std::ffi::c_char,
-            settings: *const std::ffi::c_char,
-            preferences: *const std::ffi::c_char,
-            language: *const std::ffi::c_char,
-            about: *const std::ffi::c_char,
-            quit: *const std::ffi::c_char,
-            hide: *const std::ffi::c_char,
-            hide_others: *const std::ffi::c_char,
-            show_all: *const std::ffi::c_char,
-            check_updates: *const std::ffi::c_char,
-            help: *const std::ffi::c_char,
-            release_notes: *const std::ffi::c_char,
-        );
-    }
-}
-
-/// Initializes the macOS native menu bar.
-/// Called from main.rs after eframe creates the window.
-///
-/// # Safety
-/// Contains Objective-C runtime calls. Must be called only once from the main thread.
-#[cfg(all(target_os = "macos", not(test)))]
-pub unsafe fn native_menu_setup() {
-    native_menu::katana_setup_native_menu();
-}
-
-/// Sets the macOS process name to "KatanA".
-/// Must be called at the very start of main(), BEFORE eframe creates the window,
-/// so that the Dock label shows "KatanA" instead of the binary name.
-///
-/// # Safety
-/// Contains Objective-C runtime calls. Must be called from the main thread.
-#[cfg(all(target_os = "macos", not(test)))]
-pub unsafe fn native_set_process_name() {
-    native_menu::katana_set_process_name();
-}
-
-/// Sets the macOS application icon dynamically from PNG data.
-///
-/// # Safety
-/// Contains Objective-C runtime calls. Must be called from the main thread.
-#[cfg(all(target_os = "macos", not(test)))]
-pub unsafe fn native_set_app_icon_png(png_data: *const u8, png_len: usize) {
-    native_menu::katana_set_app_icon_png(png_data, png_len as std::ffi::c_ulong);
-}
+pub use crate::native_menu::update_native_menu_strings_from_i18n;
 
 #[cfg(all(target_os = "macos", not(test)))]
-#[allow(clippy::too_many_arguments)]
-unsafe fn native_update_menu_strings(
-    file: &str,
-    open_workspace: &str,
-    save: &str,
-    settings: &str,
-    preferences: &str,
-    language: &str,
-    about: &str,
-    quit: &str,
-    hide: &str,
-    hide_others: &str,
-    show_all: &str,
-    check_updates: &str,
-    help: &str,
-    release_notes: &str,
-) {
-    let f = std::ffi::CString::new(file).unwrap_or_default();
-    let ow = std::ffi::CString::new(open_workspace).unwrap_or_default();
-    let s = std::ffi::CString::new(save).unwrap_or_default();
-    let st = std::ffi::CString::new(settings).unwrap_or_default();
-    let p = std::ffi::CString::new(preferences).unwrap_or_default();
-    let l = std::ffi::CString::new(language).unwrap_or_default();
-    let a = std::ffi::CString::new(about).unwrap_or_default();
-    let q = std::ffi::CString::new(quit).unwrap_or_default();
-    let h = std::ffi::CString::new(hide).unwrap_or_default();
-    let ho = std::ffi::CString::new(hide_others).unwrap_or_default();
-    let sa = std::ffi::CString::new(show_all).unwrap_or_default();
-    let cu = std::ffi::CString::new(check_updates).unwrap_or_default();
-    let hlp = std::ffi::CString::new(help).unwrap_or_default();
-    let rn = std::ffi::CString::new(release_notes).unwrap_or_default();
-    native_menu::katana_update_menu_strings(
-        f.as_ptr(),
-        ow.as_ptr(),
-        s.as_ptr(),
-        st.as_ptr(),
-        p.as_ptr(),
-        l.as_ptr(),
-        a.as_ptr(),
-        q.as_ptr(),
-        h.as_ptr(),
-        ho.as_ptr(),
-        sa.as_ptr(),
-        cu.as_ptr(),
-        hlp.as_ptr(),
-        rn.as_ptr(),
-    );
-}
-
-#[cfg(all(target_os = "macos", not(test)))]
-pub fn update_native_menu_strings_from_i18n() {
-    let msgs = crate::i18n::get();
-    let preferences = format!("{}…", msgs.menu.settings);
-    unsafe {
-        native_update_menu_strings(
-            &msgs.menu.file,
-            &msgs.menu.open_workspace,
-            &msgs.menu.save,
-            &msgs.menu.settings,
-            &preferences,
-            &msgs.menu.language,
-            &msgs.menu.about,
-            &msgs.menu.quit,
-            &msgs.menu.hide,
-            &msgs.menu.hide_others,
-            &msgs.menu.show_all,
-            &msgs.menu.check_updates,
-            &msgs.menu.help,
-            &msgs.menu.release_notes,
-        );
-    }
-}
-
-#[cfg(any(not(target_os = "macos"), test))]
-pub fn update_native_menu_strings_from_i18n() {}
-
-// ─────────────────────────────────────────────
-// eframe::App Implementation (egui Main Rendering Loop)
-// ─────────────────────────────────────────────
-
-use crate::shell::{KatanaApp, SIDEBAR_COLLAPSED_TOGGLE_WIDTH};
+pub use crate::native_menu::{native_menu_setup, native_set_app_icon_png, native_set_process_name};
+use crate::shell::KatanaApp;
 
 // Half-panel ratio for responsive 50/50 split.
 pub(crate) const SPLIT_HALF_RATIO: f32 = 0.5;
@@ -247,176 +82,6 @@ pub(crate) const SPLIT_HALF_RATIO: f32 = 0.5;
 /// guaranteeing the editor retains at least 30% for scrolling.
 pub(crate) const SPLIT_PANEL_MAX_RATIO: f32 = 0.7;
 pub(crate) const PREVIEW_CONTENT_PADDING: i8 = 12;
-
-impl KatanaApp {
-    const TERMS_MODAL_WIDTH: f32 = 600.0;
-    const TERMS_TITLE_SIZE: f32 = 28.0;
-    const TERMS_INNER_MARGIN: f32 = 24.0;
-    const TERMS_CONVAS_MARGIN: f32 = 16.0;
-    const TERMS_ROUNDING_LARGE: f32 = 12.0;
-    const TERMS_ROUNDING_SMALL: f32 = 8.0;
-    const TERMS_SPACING_SMALL: f32 = 8.0;
-    const TERMS_SPACING_MEDIUM: f32 = 20.0;
-    const TERMS_SPACING_XLARGE: f32 = 32.0;
-    const TERMS_BUTTON_WIDTH: f32 = 120.0;
-    const TERMS_BUTTON_HEIGHT: f32 = 40.0;
-    const TERMS_BUTTON_TEXT_SIZE: f32 = 16.0;
-    const TERMS_BUTTON_SPACING: f32 = 24.0;
-    const TERMS_SCROLL_HEIGHT_RATIO: f32 = 0.5;
-    const TERMS_CENTER_OFFSET_RATIO: f32 = 0.1;
-    const TERMS_LANG_SELECT_WIDTH: f32 = 140.0;
-
-    fn render_terms_modal(&mut self, ctx: &egui::Context, version: &str) {
-        let terms = crate::i18n::get().terms.clone();
-
-        egui::CentralPanel::default()
-            .frame(egui::Frame::central_panel(&ctx.style()).inner_margin(0.0))
-            .show(ctx, |ui| {
-                let width = ui.available_width();
-                let height = ui.available_height();
-                let content_width = width.min(Self::TERMS_MODAL_WIDTH);
-
-                ui.vertical_centered(|ui| {
-                    ui.add_space(height * Self::TERMS_CENTER_OFFSET_RATIO);
-
-                    ui.set_width(content_width);
-
-                    egui::Frame::window(ui.style())
-                        .inner_margin(Self::TERMS_INNER_MARGIN)
-                        .corner_radius(Self::TERMS_ROUNDING_LARGE)
-                        .show(ui, |ui| {
-                            ui.vertical_centered(|ui| {
-                                ui.heading(
-                                    egui::RichText::new(&terms.title)
-                                        .size(Self::TERMS_TITLE_SIZE)
-                                        .strong()
-                                        .color(ui.visuals().strong_text_color()),
-                                );
-                                ui.add_space(Self::TERMS_SPACING_SMALL);
-
-                                ui.horizontal(|ui| {
-                                    ui.label(
-                                        egui::RichText::new(crate::i18n::tf(
-                                            &terms.version_label,
-                                            &[("version", version)],
-                                        ))
-                                        .weak(),
-                                    );
-
-                                    ui.with_layout(
-                                        egui::Layout::right_to_left(egui::Align::Center),
-                                        |ui| {
-                                            let current_lang = crate::i18n::get_language();
-                                            let current_name = crate::i18n::supported_languages()
-                                                .iter()
-                                                .find(|(code, _)| *code == current_lang)
-                                                .map(|(_, name)| name.as_str())
-                                                .unwrap_or("English");
-
-                                            StyledComboBox::new("terms_lang_select", current_name)
-                                                .width(Self::TERMS_LANG_SELECT_WIDTH)
-                                                .show(ui, |ui| {
-                                                    for (code, name) in
-                                                        crate::i18n::supported_languages()
-                                                    {
-                                                        if ui
-                                                            .selectable_label(
-                                                                current_lang == *code,
-                                                                name,
-                                                            )
-                                                            .clicked()
-                                                        {
-                                                            self.pending_action =
-                                                                AppAction::ChangeLanguage(
-                                                                    code.clone(),
-                                                                );
-                                                        }
-                                                    }
-                                                });
-                                        },
-                                    );
-                                });
-
-                                ui.add_space(Self::TERMS_SPACING_MEDIUM);
-                                ui.separator();
-                                ui.add_space(Self::TERMS_SPACING_MEDIUM);
-
-                                egui::Frame::canvas(ui.style())
-                                    .inner_margin(Self::TERMS_CONVAS_MARGIN)
-                                    .corner_radius(Self::TERMS_ROUNDING_SMALL)
-                                    .show(ui, |ui| {
-                                        ui.set_min_height(
-                                            ui.available_height() * Self::TERMS_SCROLL_HEIGHT_RATIO,
-                                        );
-                                        egui::ScrollArea::vertical()
-                                            .max_height(
-                                                ui.available_height()
-                                                    * Self::TERMS_SCROLL_HEIGHT_RATIO,
-                                            )
-                                            .show(ui, |ui| {
-                                                ui.add(egui::Label::new(&terms.content).wrap());
-                                            });
-                                    });
-
-                                ui.add_space(Self::TERMS_SPACING_XLARGE);
-
-                                ui.horizontal(|ui| {
-                                    let total_buttons_width =
-                                        Self::TERMS_BUTTON_WIDTH * 2.0 + Self::TERMS_BUTTON_SPACING;
-                                    let available = ui.available_width();
-                                    let outer_spacing = (available - total_buttons_width) / 2.0;
-
-                                    if outer_spacing > 0.0 {
-                                        ui.add_space(outer_spacing);
-                                    }
-
-                                    let accept_btn = egui::Button::new(
-                                        egui::RichText::new(&terms.accept)
-                                            .strong()
-                                            .size(Self::TERMS_BUTTON_TEXT_SIZE),
-                                    )
-                                    .min_size(egui::vec2(
-                                        Self::TERMS_BUTTON_WIDTH,
-                                        Self::TERMS_BUTTON_HEIGHT,
-                                    ))
-                                    .corner_radius(Self::TERMS_ROUNDING_SMALL);
-
-                                    if ui
-                                        .add(accept_btn)
-                                        .on_hover_cursor(egui::CursorIcon::PointingHand)
-                                        .clicked()
-                                    {
-                                        self.pending_action =
-                                            AppAction::AcceptTerms(version.to_string());
-                                    }
-
-                                    ui.add_space(Self::TERMS_BUTTON_SPACING);
-
-                                    let decline_btn = egui::Button::new(
-                                        egui::RichText::new(&terms.decline)
-                                            .size(Self::TERMS_BUTTON_TEXT_SIZE),
-                                    )
-                                    .min_size(egui::vec2(
-                                        Self::TERMS_BUTTON_WIDTH,
-                                        Self::TERMS_BUTTON_HEIGHT,
-                                    ))
-                                    .corner_radius(Self::TERMS_ROUNDING_SMALL);
-
-                                    if ui
-                                        .add(decline_btn)
-                                        .on_hover_cursor(egui::CursorIcon::PointingHand)
-                                        .clicked()
-                                    {
-                                        self.pending_action = AppAction::DeclineTerms;
-                                    }
-                                });
-                                ui.add_space(Self::TERMS_SPACING_MEDIUM);
-                            });
-                        });
-                });
-            });
-    }
-}
 
 impl eframe::App for KatanaApp {
     fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
@@ -542,62 +207,10 @@ impl eframe::App for KatanaApp {
         self.poll_export(ctx);
 
         // macOS: Poll actions from the native menu.
-        #[cfg(target_os = "macos")]
-        {
-            let action = unsafe { native_menu::katana_poll_menu_action() };
-            match action {
-                native_menu::TAG_OPEN_WORKSPACE => {
-                    if let Some(path) = open_folder_dialog() {
-                        self.pending_action = AppAction::OpenWorkspace(path);
-                    }
-                }
-                native_menu::TAG_SAVE => {
-                    self.pending_action = AppAction::SaveDocument;
-                }
-                native_menu::TAG_LANG_EN => {
-                    self.pending_action = AppAction::ChangeLanguage("en".to_string());
-                }
-                native_menu::TAG_LANG_JA => {
-                    self.pending_action = AppAction::ChangeLanguage("ja".to_string());
-                }
-                native_menu::TAG_LANG_ZH_CN => {
-                    self.pending_action = AppAction::ChangeLanguage("zh-CN".to_string());
-                }
-                native_menu::TAG_LANG_ZH_TW => {
-                    self.pending_action = AppAction::ChangeLanguage("zh-TW".to_string());
-                }
-                native_menu::TAG_LANG_KO => {
-                    self.pending_action = AppAction::ChangeLanguage("ko".to_string());
-                }
-                native_menu::TAG_LANG_PT => {
-                    self.pending_action = AppAction::ChangeLanguage("pt".to_string());
-                }
-                native_menu::TAG_LANG_FR => {
-                    self.pending_action = AppAction::ChangeLanguage("fr".to_string());
-                }
-                native_menu::TAG_LANG_DE => {
-                    self.pending_action = AppAction::ChangeLanguage("de".to_string());
-                }
-                native_menu::TAG_LANG_ES => {
-                    self.pending_action = AppAction::ChangeLanguage("es".to_string());
-                }
-                native_menu::TAG_LANG_IT => {
-                    self.pending_action = AppAction::ChangeLanguage("it".to_string());
-                }
-                native_menu::TAG_ABOUT => {
-                    self.show_about = !self.show_about;
-                }
-                native_menu::TAG_CHECK_UPDATES => {
-                    self.pending_action = AppAction::CheckForUpdates;
-                }
-                native_menu::TAG_RELEASE_NOTES => {
-                    self.pending_action = AppAction::ShowReleaseNotes;
-                }
-                native_menu::TAG_SETTINGS => {
-                    self.pending_action = AppAction::ToggleSettings;
-                }
-                _ => {}
-            }
+        let native_action =
+            crate::native_menu::poll_native_menu(&mut self.show_about, open_folder_dialog);
+        if !matches!(native_action, AppAction::None) {
+            self.pending_action = native_action;
         }
 
         let action = self.take_action();
@@ -615,223 +228,18 @@ impl eframe::App for KatanaApp {
                 .terms_accepted_version
                 .as_ref();
             if accepted_ver != Some(&terms_ver) {
-                self.render_terms_modal(ctx, &terms_ver);
+                crate::views::modals::terms::render_terms_modal(
+                    ctx,
+                    &terms_ver,
+                    &mut self.pending_action,
+                );
                 return;
             }
         }
 
         if !splash_is_opaque {
-            // On macOS, the egui menu is hidden because the native menu bar is used.
-            crate::views::top_bar::render_menu_bar(ctx, &mut self.state, &mut self.pending_action);
-            let export_filenames: Vec<String> = self
-                .export_tasks
-                .iter()
-                .map(|t| t.filename.clone())
-                .collect();
-            crate::views::top_bar::render_status_bar(ctx, &self.state, &export_filenames);
-
-            // Reflect the file name in the window title
-            let ws_root_for_title = self.state.workspace.data.as_ref().map(|ws| ws.root.clone());
-            let title_text = match self.state.active_document() {
-                Some(doc) => {
-                    let fname = doc.file_name().unwrap_or("");
-                    let rel = relative_full_path(&doc.path, ws_root_for_title.as_deref());
-                    crate::shell_logic::format_window_title(
-                        fname,
-                        &rel,
-                        &crate::i18n::get().menu.release_notes,
-                    )
-                }
-                None => "KatanA".to_string(),
-            };
-            if self.state.layout.last_window_title != title_text {
-                ctx.send_viewport_cmd(egui::ViewportCommand::Title(title_text.clone()));
-                self.state.layout.last_window_title = title_text.clone();
-            }
-
-            // In-app title bar
-            egui::TopBottomPanel::top("app_title_bar").show(ctx, |ui| {
-                ui.horizontal(|ui| {
-                    ui.centered_and_justified(|ui| {
-                        let title_color =
-                            theme_bridge::rgb_to_color32(theme_colors.system.title_bar_text);
-                        ui.label(egui::RichText::new(&title_text).small().color(title_color));
-                    });
-                });
-            });
-
-            // Show the collapse toggle button even when the workspace is hidden.
-            if !self.state.layout.show_workspace {
-                egui::SidePanel::left("workspace_collapsed")
-                    .resizable(false)
-                    .exact_width(SIDEBAR_COLLAPSED_TOGGLE_WIDTH)
-                    .show(ctx, |ui| {
-                        ui.vertical_centered(|ui| {
-                            if ui
-                                .add(egui::Button::image(
-                                    crate::Icon::ChevronRight
-                                        .ui_image(ui, crate::icon::IconSize::Medium),
-                                ))
-                                .on_hover_text(crate::i18n::get().workspace.workspace_title.clone())
-                                .clicked()
-                            {
-                                self.state.layout.show_workspace = true;
-                            }
-                        });
-                    });
-            } else {
-                crate::views::panels::workspace::render_workspace_panel(
-                    ctx,
-                    &mut self.state,
-                    &mut self.pending_action,
-                );
-            }
-
-            // Tab row + breadcrumbs + view mode row
-            egui::TopBottomPanel::top("tab_toolbar").show(ctx, |ui| {
-                crate::views::top_bar::render_tab_bar(ui, &mut self.state, &mut self.pending_action);
-                let active_doc_props = self.state.active_document();
-                if let Some(doc) = active_doc_props {
-                    let d_path = doc.path.to_string_lossy();
-                    let is_changelog = d_path.starts_with("Katana://ChangeLog");
-
-                    if !is_changelog {
-                        let doc_path = doc.path.clone();
-                        let ws_root = self.state.workspace.data.as_ref().map(|ws| ws.root.clone());
-                        let rel = relative_full_path(&doc_path, ws_root.as_deref());
-                        let mut breadcrumb_action = None;
-                        ui.horizontal(|ui| {
-                            let segments: Vec<&str> = rel.split('/').collect();
-                            let mut current_path = ws_root.clone().unwrap_or_default();
-                            for (i, seg) in segments.iter().enumerate() {
-                                if i > 0 {
-                                    const CHEVRON_ICON_SIZE: f32 = 10.0;
-                                    ui.add(
-                                        egui::Image::new(crate::Icon::ChevronRight.uri())
-                                            .tint(ui.visuals().text_color())
-                                            .max_height(CHEVRON_ICON_SIZE),
-                                    );
-                                }
-
-                                if ws_root.is_none() {
-                                    ui.label(egui::RichText::new(*seg).small());
-                                    continue;
-                                }
-
-                                current_path = current_path.join(seg);
-                                let is_last = i == segments.len() - 1;
-
-                                // The doc_path_clone is only needed for the original `render_tree_entry` logic,
-                                // which is being replaced.
-                                // let doc_path_clone = doc_path.clone();
-
-                                if is_last {
-                                    // Last segment is the file itself, just render it as text
-                                    ui.add(
-                                        egui::Label::new(egui::RichText::new(*seg).small())
-                                            .sense(egui::Sense::hover()),
-                                    );
-                                } else {
-                                    // Dropdown mini-workspace for Breadcrumbs
-                                    ui.menu_button(egui::RichText::new(*seg).small(), |ui| {
-                                        let mut ctx_action = crate::app_state::AppAction::None;
-
-                                        if let Some(ws) = &self.state.workspace.data {
-                                            if let Some(
-                                                katana_core::workspace::TreeEntry::Directory {
-                                                    children,
-                                                    ..
-                                                },
-                                            ) = crate::views::panels::tree::find_node_in_tree(&ws.tree, &current_path)
-                                            {
-                                                crate::views::panels::workspace::render_breadcrumb_menu(
-                                                    ui,
-                                                    children,
-                                                    &mut ctx_action,
-                                                );
-                                            }
-                                        }
-
-                                        if !matches!(ctx_action, crate::app_state::AppAction::None)
-                                        {
-                                            breadcrumb_action = Some(ctx_action);
-                                            ui.close();
-                                        }
-                                    });
-                                }
-                            }
-                        });
-                        if let Some(a) = breadcrumb_action {
-                            self.pending_action = a;
-                        }
-                    } // End if !is_changelog
-                    crate::views::top_bar::render_view_mode_bar(ui, &mut self.state, &mut self.pending_action);
-                }
-            });
-
-            let mut download_req: Option<DownloadRequest> = None;
-            let current_mode = self.state.active_view_mode();
-            let is_split = current_mode == ViewMode::Split;
-            let mut is_changelog_tab = false;
-
-            if let Some(doc) = self.state.active_document() {
-                if doc.path.to_string_lossy().starts_with("Katana://ChangeLog") {
-                    is_changelog_tab = true;
-                }
-            }
-
-            if self.state.layout.show_toc
-                && self.state.config.settings.settings().layout.toc_visible
-            {
-                if let Some(doc) = self.state.active_document() {
-                    if let Some(preview) = self.tab_previews.iter_mut().find(|p| p.path == doc.path)
-                    {
-                        crate::views::panels::toc::render_toc_panel(
-                            ctx,
-                            &mut preview.pane,
-                            &self.state,
-                        );
-                    }
-                }
-            }
-
-            if is_changelog_tab {
-                egui::CentralPanel::default().show(ctx, |ui| {
-                    crate::changelog::render_release_notes_tab(
-                        ui,
-                        &self.changelog_sections,
-                        self.changelog_rx.is_some(),
-                    );
-                });
-            } else {
-                if is_split {
-                    let split_dir = self.state.active_split_direction();
-                    let pane_order = self.state.active_pane_order();
-                    download_req = crate::views::layout::split::render_split_mode(
-                        ctx, self, split_dir, pane_order,
-                    );
-                }
-
-                if !is_split {
-                    egui::CentralPanel::default()
-                        .frame(egui::Frame::central_panel(&ctx.style()).inner_margin(0.0))
-                        .show(ctx, |ui| match current_mode {
-                            ViewMode::CodeOnly => {
-                                crate::views::panels::editor::render_editor_content(
-                                    ui,
-                                    &mut self.state,
-                                    &mut self.pending_action,
-                                    false,
-                                );
-                            }
-                            ViewMode::PreviewOnly => {
-                                crate::views::layout::split::render_preview_only(ui, self);
-                            }
-                            ViewMode::Split => {}
-                        });
-                }
-            }
-
+            let download_req =
+                crate::views::app_frame::render_main_panels(ctx, self, &theme_colors);
             if let Some(req) = download_req {
                 self.start_download(req);
             }
@@ -948,151 +356,10 @@ impl eframe::App for KatanaApp {
         // --- Splash Screen Overlay ---
         if let Some(start) = self.splash_start {
             let elapsed = start.elapsed().as_secs_f32();
-            let opacity = crate::shell_logic::calculate_splash_opacity(elapsed);
-            let any_pressed = ctx.input(|i| i.key_pressed(egui::Key::Escape));
-
-            if opacity <= 0.0 || any_pressed {
+            let dismissed =
+                crate::views::splash::render_splash_overlay(ctx, elapsed, self.about_icon.as_ref());
+            if dismissed {
                 self.splash_start = None;
-            } else {
-                egui::Area::new(egui::Id::new("splash_screen_area"))
-                    .order(egui::Order::Foreground)
-                    .interactable(true) // Consume interactions directly falling through
-                    .show(ctx, |ui| {
-                        const SPLASH_BG_DARK: u8 = 30;
-                        const SPLASH_BG_LIGHT: u8 = 240;
-                        const SPLASH_ICON_SIZE: f32 = 128.0;
-                        const SPLASH_ICON_SPACING: f32 = 16.0;
-                        const SPLASH_HEADING_SIZE: f32 = 32.0;
-                        const SPLASH_HEADING_SPACING: f32 = 8.0;
-                        const SPLASH_VERSION_SIZE: f32 = 16.0;
-                        const SPLASH_PROGRESS_SPACING: f32 = 24.0;
-                        const SPLASH_PROGRESS_WIDTH: f32 = 240.0;
-                        const SPLASH_PROGRESS_PHASE1: f32 = 0.25;
-                        const SPLASH_PROGRESS_PHASE2: f32 = 0.6;
-                        const SPLASH_PROGRESS_PHASE3: f32 = 0.95;
-                        const SPLASH_PROGRESS_TEXT_SIZE: f32 = 12.0;
-                        const SPLASH_PROGRESS_TEXT_DIM: f32 = 0.7;
-                        const SPLASH_PROGRESS_BAR_MARGIN: f32 = 4.0;
-                        const SPLASH_PROGRESS_BG_LIGHT: u8 = 100;
-                        const SPLASH_PROGRESS_BG_DARK: u8 = 200;
-
-                        let is_dark = ctx.style().visuals.dark_mode;
-                        #[allow(deprecated)]
-                        let content_rect = ctx.screen_rect();
-                        let bg_color = if is_dark {
-                            crate::theme_bridge::from_rgb(
-                                SPLASH_BG_DARK,
-                                SPLASH_BG_DARK,
-                                SPLASH_BG_DARK,
-                            )
-                        } else {
-                            crate::theme_bridge::from_rgb(
-                                SPLASH_BG_LIGHT,
-                                SPLASH_BG_LIGHT,
-                                SPLASH_BG_LIGHT,
-                            )
-                        };
-                        let fill_color = bg_color.gamma_multiply(opacity);
-                        ui.painter().rect_filled(content_rect, 1.0, fill_color);
-
-                        let text_color = if is_dark {
-                            crate::theme_bridge::WHITE
-                        } else {
-                            crate::theme_bridge::TRANSPARENT
-                        }
-                        .gamma_multiply(opacity);
-
-                        // Calculate total content height to vertically center the splash content
-                        const SPLASH_CONTENT_HEIGHT: f32 = SPLASH_ICON_SIZE
-                            + SPLASH_ICON_SPACING
-                            + SPLASH_HEADING_SIZE
-                            + SPLASH_HEADING_SPACING
-                            + SPLASH_VERSION_SIZE
-                            + SPLASH_PROGRESS_SPACING
-                            + SPLASH_PROGRESS_TEXT_SIZE
-                            + SPLASH_PROGRESS_BAR_MARGIN
-                            + SPLASH_PROGRESS_SPACING; // approximate total height
-
-                        let center = content_rect.center();
-                        let centered_rect = egui::Rect::from_center_size(
-                            center,
-                            egui::vec2(content_rect.width(), SPLASH_CONTENT_HEIGHT),
-                        );
-
-                        ui.scope_builder(egui::UiBuilder::new().max_rect(centered_rect), |ui| {
-                            ui.vertical_centered(|ui| {
-                                if let Some(tex) = &self.about_icon {
-                                    ui.image(egui::load::SizedTexture::new(
-                                        tex.id(),
-                                        egui::vec2(SPLASH_ICON_SIZE, SPLASH_ICON_SIZE),
-                                    ));
-                                    ui.add_space(SPLASH_ICON_SPACING);
-                                }
-                                let heading =
-                                    egui::RichText::new(crate::about_info::APP_DISPLAY_NAME)
-                                        .strong()
-                                        .size(SPLASH_HEADING_SIZE)
-                                        .color(text_color);
-                                ui.label(heading);
-
-                                ui.add_space(SPLASH_HEADING_SPACING);
-
-                                let version_str = format!("Version {}", env!("CARGO_PKG_VERSION"));
-                                let version = egui::RichText::new(version_str)
-                                    .size(SPLASH_VERSION_SIZE)
-                                    .color(text_color);
-                                ui.label(version);
-
-                                ui.add_space(SPLASH_PROGRESS_SPACING);
-                                let progress =
-                                    crate::shell_logic::calculate_splash_progress(elapsed);
-
-                                let progress_text = if progress < SPLASH_PROGRESS_PHASE1 {
-                                    "Initializing Katana engine..."
-                                } else if progress < SPLASH_PROGRESS_PHASE2 {
-                                    "Parsing workspace structure..."
-                                } else if progress < SPLASH_PROGRESS_PHASE3 {
-                                    "Increasing context size... w"
-                                } else {
-                                    "Ready."
-                                };
-
-                                ui.label(
-                                    egui::RichText::new(progress_text)
-                                        .size(SPLASH_PROGRESS_TEXT_SIZE)
-                                        .color(text_color.gamma_multiply(SPLASH_PROGRESS_TEXT_DIM)),
-                                );
-                                ui.add_space(SPLASH_PROGRESS_BAR_MARGIN);
-                                let progress_bar = egui::ProgressBar::new(progress)
-                                    .desired_width(SPLASH_PROGRESS_WIDTH)
-                                    .show_percentage();
-
-                                // Add a little visual flair to the progress bar by tinting it based on the theme
-                                if !is_dark {
-                                    ui.visuals_mut().selection.bg_fill =
-                                        crate::theme_bridge::from_rgb(
-                                            SPLASH_PROGRESS_BG_LIGHT,
-                                            SPLASH_PROGRESS_BG_LIGHT,
-                                            SPLASH_PROGRESS_BG_LIGHT,
-                                        )
-                                        .gamma_multiply(opacity);
-                                } else {
-                                    ui.visuals_mut().selection.bg_fill =
-                                        crate::theme_bridge::from_rgb(
-                                            SPLASH_PROGRESS_BG_DARK,
-                                            SPLASH_PROGRESS_BG_DARK,
-                                            SPLASH_PROGRESS_BG_DARK,
-                                        )
-                                        .gamma_multiply(opacity);
-                                }
-                                ui.add(progress_bar);
-                            });
-                        });
-                    });
-                // Animate splash screen (fade in/out, progress text)
-                ctx.request_repaint_after(std::time::Duration::from_millis(
-                    SPLASH_REPAINT_INTERVAL_MS,
-                ));
             }
         }
     }
