+++ crates/katana-ui/src/views/mod.rs
@@ -1,4 +1,6 @@
+pub mod app_frame;
 pub mod layout;
 pub mod modals;
 pub mod panels;
+pub mod splash;
 pub mod top_bar;
