+++ openspec/changes/v0-8-4-rust-code-refactoring/tasks.md
@@ -302,6 +302,17 @@ Tasks Grouped by ## = Adhere unconditionally to the branching standard defined i
 
 > 6-A 〜 6-D の「構造分割」が終わった後、UI を単なる free function の寄せ集めではなく、再利用可能で再現性の高い component 境界に揃える。
 
+#### Definition of Ready (DoR)
+
+> **6-Eに着手する前に、以下がすべて満たされていること。**
+
+- [ ] `shell_ui.rs` のロジック行数（テスト除外）が200行以下であること
+  - 現状: 1,107行（`impl eframe::App::update()` 686行 + `impl KatanaApp` 170行 + ユーティリティ/定数 251行）
+  - 対処: `update()` 内のオーケストレーションロジックを `views/app_frame.rs` 等に抽出、`render_terms_modal` を `views/modals/terms.rs` に移動、定数・ユーティリティを使用元モジュールに分散
+- [ ] `shell.rs` のロジック行数（テスト除外）が200行以下であること（現状329行 → 定数/型を使用元に移動）
+- [ ] `views/modals/` の全モーダルが `shell_ui.rs` のローカルコピーではなく正規の実装であること（✅ 完了済み）
+- [ ] `make check` がパス
+
 - [ ] 6.24 `ui/menu`, `ui/header`, `ui/status_bar`, `ui/workspace`, `ui/tab_bar`, `ui/modals` を `struct + impl show() -> Response` パターンへ統一
 - [ ] 6.25 `settings/`, `preview/`, `widgets/` の各UIを props + typed response を持つ自己完結コンポーネントへ統一
 - [ ] 6.26 親子 UI 間の依存を最小 props + typed response に整理し、巨大な `AppState` / `KatanaApp` の横流しを段階的に排除
